requirements.txt
langchain>=0.1.0
langchain-openai
openai
rich
python-dotenv
streamlit